レイアウト崩れを対処
